+++ tencentcloud/mgobe/v20201014/client.go
@@ -63,6 +63,10 @@ func NewChangeRoomPlayerProfileResponse() (response *ChangeRoomPlayerProfileResp
 }
 
 // ChangeRoomPlayerProfile
+// 此接口无法使用，游戏联机对战引擎MGOBE已于6.1正式下架，感谢您的支持
+//
+// 
+//
 // 修改房间玩家自定义属性
 //
 // 可能返回的错误码:
@@ -89,6 +93,10 @@ func (c *Client) ChangeRoomPlayerProfile(request *ChangeRoomPlayerProfileRequest
 }
 
 // ChangeRoomPlayerProfile
+// 此接口无法使用，游戏联机对战引擎MGOBE已于6.1正式下架，感谢您的支持
+//
+// 
+//
 // 修改房间玩家自定义属性
 //
 // 可能返回的错误码:
@@ -144,6 +152,10 @@ func NewChangeRoomPlayerStatusResponse() (response *ChangeRoomPlayerStatusRespon
 }
 
 // ChangeRoomPlayerStatus
+// 此接口无法使用，游戏联机对战引擎MGOBE已于6.1正式下架，感谢您的支持
+//
+// 
+//
 // 修改玩家自定义状态
 //
 // 可能返回的错误码:
@@ -462,6 +474,10 @@ func (c *Client) ChangeRoomPlayerStatus(request *ChangeRoomPlayerStatusRequest) 
 }
 
 // ChangeRoomPlayerStatus
+// 此接口无法使用，游戏联机对战引擎MGOBE已于6.1正式下架，感谢您的支持
+//
+// 
+//
 // 修改玩家自定义状态
 //
 // 可能返回的错误码:
@@ -809,6 +825,10 @@ func NewDescribePlayerResponse() (response *DescribePlayerResponse) {
 }
 
 // DescribePlayer
+// 此接口无法使用，游戏联机对战引擎MGOBE已于6.1正式下架，感谢您的支持
+//
+// 
+//
 // 该接口用于查询玩家信息。支持两种用法，当OpenId不传的时候，PlayerId必传，传入PlayerId可以查询当前PlayerId的玩家信息，当OpenId传入的时候，PlayerId可不传，按照OpenId查询玩家信息。
 //
 // 可能返回的错误码:
@@ -1128,6 +1148,10 @@ func (c *Client) DescribePlayer(request *DescribePlayerRequest) (response *Descr
 }
 
 // DescribePlayer
+// 此接口无法使用，游戏联机对战引擎MGOBE已于6.1正式下架，感谢您的支持
+//
+// 
+//
 // 该接口用于查询玩家信息。支持两种用法，当OpenId不传的时候，PlayerId必传，传入PlayerId可以查询当前PlayerId的玩家信息，当OpenId传入的时候，PlayerId可不传，按照OpenId查询玩家信息。
 //
 // 可能返回的错误码:
@@ -1476,6 +1500,10 @@ func NewDescribeRoomResponse() (response *DescribeRoomResponse) {
 }
 
 // DescribeRoom
+// 此接口无法使用，游戏联机对战引擎MGOBE已于6.1正式下架，感谢您的支持
+//
+// 
+//
 // 该接口用于查询房间信息。支持两种用法，当房间Id不传的时候，玩家Id必传，传入玩家Id可以查询当前玩家所在的房间信息，当房间Id传入的时候，玩家Id可不传，按照房间Id查询房间信息。
 //
 // 可能返回的错误码:
@@ -1795,6 +1823,10 @@ func (c *Client) DescribeRoom(request *DescribeRoomRequest) (response *DescribeR
 }
 
 // DescribeRoom
+// 此接口无法使用，游戏联机对战引擎MGOBE已于6.1正式下架，感谢您的支持
+//
+// 
+//
 // 该接口用于查询房间信息。支持两种用法，当房间Id不传的时候，玩家Id必传，传入玩家Id可以查询当前玩家所在的房间信息，当房间Id传入的时候，玩家Id可不传，按照房间Id查询房间信息。
 //
 // 可能返回的错误码:
@@ -2143,6 +2175,10 @@ func NewDismissRoomResponse() (response *DismissRoomResponse) {
 }
 
 // DismissRoom
+// 此接口无法使用，游戏联机对战引擎MGOBE已于6.1正式下架，感谢您的支持
+//
+// 
+//
 // 通过game_id、room_id解散房间
 //
 // 可能返回的错误码:
@@ -2461,6 +2497,10 @@ func (c *Client) DismissRoom(request *DismissRoomRequest) (response *DismissRoom
 }
 
 // DismissRoom
+// 此接口无法使用，游戏联机对战引擎MGOBE已于6.1正式下架，感谢您的支持
+//
+// 
+//
 // 通过game_id、room_id解散房间
 //
 // 可能返回的错误码:
@@ -2808,6 +2848,10 @@ func NewModifyRoomResponse() (response *ModifyRoomResponse) {
 }
 
 // ModifyRoom
+// 此接口无法使用，游戏联机对战引擎MGOBE已于6.1正式下架，感谢您的支持
+//
+// 
+//
 // 修改房间
 //
 // 可能返回的错误码:
@@ -3126,6 +3170,10 @@ func (c *Client) ModifyRoom(request *ModifyRoomRequest) (response *ModifyRoomRes
 }
 
 // ModifyRoom
+// 此接口无法使用，游戏联机对战引擎MGOBE已于6.1正式下架，感谢您的支持
+//
+// 
+//
 // 修改房间
 //
 // 可能返回的错误码:
@@ -3473,6 +3521,10 @@ func NewRemoveRoomPlayerResponse() (response *RemoveRoomPlayerResponse) {
 }
 
 // RemoveRoomPlayer
+// 此接口无法使用，游戏联机对战引擎MGOBE已于6.1正式下架，感谢您的支持
+//
+// 
+//
 // 踢出房间玩家
 //
 // 可能返回的错误码:
@@ -3791,6 +3843,10 @@ func (c *Client) RemoveRoomPlayer(request *RemoveRoomPlayerRequest) (response *R
 }
 
 // RemoveRoomPlayer
+// 此接口无法使用，游戏联机对战引擎MGOBE已于6.1正式下架，感谢您的支持
+//
+// 
+//
 // 踢出房间玩家
 //
 // 可能返回的错误码: